requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
filelock==3.18.0
flake8==7.2.0
frozenlist==1.6.0
fsspec==2025.5.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.1
pycodestyle==2.13.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pytest==8.4.0
pytest-asyncio==1.0.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
setuptools==80.9.0
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
